+++ packages/portfolio/src/pages/assets/assetsPage/assets.ui.desktop.tsx
@@ -70,7 +70,7 @@ const TotalValueInfo: React.FC<
         >
           {totalValue ?? "--"}
         </Text.numeral>
-        <Text as="div" weight="bold">
+        <Text as="div" weight="bold" className="oui-leading-6">
           USDC
         </Text>
       </Flex>