+++ packages/hooks/CHANGELOG.md
@@ -1,5 +1,69 @@
 # @orderly.network/hooks
 
+## 2.12.3
+
+### Patch Changes
+
+- publish
+- 528e3f4: publish
+- Updated dependencies
+- Updated dependencies [528e3f4]
+  - @orderly.network/core@2.12.3
+  - @orderly.network/default-evm-adapter@2.12.3
+  - @orderly.network/default-solana-adapter@2.12.3
+  - @orderly.network/net@2.12.3
+  - @orderly.network/perp@4.12.3
+  - @orderly.network/types@2.12.3
+  - @orderly.network/utils@2.12.3
+  - @orderly.network/web3-provider-ethers@2.12.3
+
+## 2.12.3-alpha.0
+
+### Patch Changes
+
+- publish
+- Updated dependencies
+  - @orderly.network/core@2.12.3-alpha.0
+  - @orderly.network/default-evm-adapter@2.12.3-alpha.0
+  - @orderly.network/default-solana-adapter@2.12.3-alpha.0
+  - @orderly.network/net@2.12.3-alpha.0
+  - @orderly.network/perp@4.12.3-alpha.0
+  - @orderly.network/types@2.12.3-alpha.0
+  - @orderly.network/utils@2.12.3-alpha.0
+  - @orderly.network/web3-provider-ethers@2.12.3-alpha.0
+
+## 2.12.2
+
+### Patch Changes
+
+- publish
+- 06007f7: publish
+- Updated dependencies
+- Updated dependencies [06007f7]
+  - @orderly.network/core@2.12.2
+  - @orderly.network/default-evm-adapter@2.12.2
+  - @orderly.network/default-solana-adapter@2.12.2
+  - @orderly.network/net@2.12.2
+  - @orderly.network/perp@4.12.2
+  - @orderly.network/types@2.12.2
+  - @orderly.network/utils@2.12.2
+  - @orderly.network/web3-provider-ethers@2.12.2
+
+## 2.12.2-alpha.0
+
+### Patch Changes
+
+- publish
+- Updated dependencies
+  - @orderly.network/core@2.12.2-alpha.0
+  - @orderly.network/default-evm-adapter@2.12.2-alpha.0
+  - @orderly.network/default-solana-adapter@2.12.2-alpha.0
+  - @orderly.network/net@2.12.2-alpha.0
+  - @orderly.network/perp@4.12.2-alpha.0
+  - @orderly.network/types@2.12.2-alpha.0
+  - @orderly.network/utils@2.12.2-alpha.0
+  - @orderly.network/web3-provider-ethers@2.12.2-alpha.0
+
 ## 2.12.1
 
 ### Patch Changes