+++ packages/tsconfig/CHANGELOG.md
@@ -1,5 +1,31 @@
 # tsconfig
 
+## 0.15.3
+
+### Patch Changes
+
+- publish
+- 528e3f4: publish
+
+## 0.15.3-alpha.0
+
+### Patch Changes
+
+- publish
+
+## 0.15.2
+
+### Patch Changes
+
+- publish
+- 06007f7: publish
+
+## 0.15.2-alpha.0
+
+### Patch Changes
+
+- publish
+
 ## 0.15.1
 
 ### Patch Changes